Add room facets and colorize facet markers
Export and import find-facet-in-room and room-facets; add facets field
to room objects and populate facet data in initialize-world. Add
find-facet-in-room helper. Add colorize-facets to render [facet] markers
with color and use it in send-room-overview. Extend "look" handling to
examine room facets as well as players and items. Tweak ASCII map layout
and print the map on room entry.